+++ tests/features/test_forwardref.py
@@ -1,21 +1,20 @@
+import pytest
+
 from ididi.graph import DependencyGraph
 
 dag = DependencyGraph()
 
 
-@dag.node
 class ServiceA:
     def __init__(self, b: "ServiceB"):  # Forward reference to ServiceB
         self.b = b
 
 
-@dag.node
 class ServiceB:
     def __init__(self, c: "ServiceC"):
         self.c = c
 
 
-@dag.node
 class ServiceC:
     def __init__(self):
         pass
@@ -30,6 +29,3 @@ def test_forward_reference():
     assert isinstance(instance, ServiceA)
     assert isinstance(instance.b, ServiceB)
     assert isinstance(instance.b.c, ServiceC)
-
-
-# @pytest.mark.skip